Refactor file upload utility functions
